Add updateLists method to keep contacts list intact
Sortable will move contact nodes around but we have
to update our contact list and contactsByCategory
object up to date manually. The problem is, if we do this
manually, we end up with doubles of contacts when we
move them. We need to have our list and lookup obj
up to date at all times, and the only way to achieve this
seems to be to reset the obj after each update.