Fix Bug with Contacts
- noticed hasContacts check was failing after moving
a newly created contact from one category to another.
I was pushing the wrong contact object to the contacts
array and failed to use loose comparison when updating
a contact in the list.
- fix bug where contactsByCategory wouldn't be initialized
with any categories that didn't have contacts in them.
I was creating this object by grouping contacts =\
This caused a bug when trying to push to an existing category that didn't have any contacts in it on
beforeMount